--- app/helpers/home_helper.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-module HomeHelper
-end
