--- app/helpers/news_helper.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module NewsHelper
-  def admin_avatar_url
-    url_for AdminUser.first.avatar
-  end
-end
