+++ config/puma.rb
@@ -39,6 +39,6 @@ pidfile ENV.fetch('PIDFILE') { 'tmp/pids/server.pid' }
 # Allow puma to be restarted by `rails restart` command.
 plugin :tmp_restart
 
-# app_root = File.expand_path('../..', __FILE__)
-# bind "unix://#{app_root}/tmp/sockets/puma.sock"
-# stdout_redirect "#{app_root}/log/puma.stdout.log", "#{app_root}/log/puma.stderr.log", true
+app_root = File.expand_path('../..', __FILE__)
+bind "unix://#{app_root}/tmp/sockets/puma.sock"
+stdout_redirect "#{app_root}/log/puma.stdout.log", "#{app_root}/log/puma.stderr.log", true
