+++ spec/helpers/active_helper_spec.rb
@@ -1,15 +1,14 @@
 require 'rails_helper'
 
 # Specs in this file have access to a helper object that includes
-# the NewsHelper. For example:
+# the ActiveHelper. For example:
 #
-# describe NewsHelper do
+# describe ActiveHelper do
 #   describe "string concat" do
 #     it "concats two strings with spaces" do
 #       expect(helper.concat_strings("this","that")).to eq("this that")
 #     end
 #   end
 # end
-RSpec.describe NewsHelper, type: :helper do
-  pending "add some examples to (or delete) #{__FILE__}"
+RSpec.describe ActiveHelper, type: :helper do
 end
